Update the bedroom lights to make more sense.
At a high level, leave the power to the ceiling lights on full time.
We still want to configure the relays, but leave them on; treating them as relays rather than leds.

Set up a group for ceiling lights so we can more clearly manipulate them as a group (and not touch the "all bedroom lights" which would touch night stand lights too).

Remove unnecessary or even harmful automation.